Archetypes provide default front matter
This one is for generic material.

You can use this by issueing

`hugo new "post/2015/06/29/Something with spaces in it even.md"`

And the folder structure will be created `content/post/2015/06/29/` with
the filename.

Now, we do need to have some care, due to Windows and Posix systems
acting differently with files with spaces.  I would perhaps generally
use dashes between words, but this does work.